# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentGuard Core Engine",
    "description": "Framework-Agnostic Runtime Governance Layer for Autonomous AI Agents",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/hitl": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "List Hitl Requests",
        "description": "List HITL requests for the caller's organization with lazy expiration.",
        "operationId": "list_hitl_requests_api_v1_hitl_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (e.g. PENDING, APPROVED, DENIED, EXPIRED)",
              "title": "Status"
            },
            "description": "Filter by status (e.g. PENDING, APPROVED, DENIED, EXPIRED)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page limit",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page limit"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HITLRequestListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{hitl_id}": {
      "get": {
        "tags": [
          "hitl"
        ],
        "summary": "Get Hitl Request",
        "description": "Retrieve a single HITL request by ID within the caller's organization.",
        "operationId": "get_hitl_request_api_v1_hitl__hitl_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hitl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hitl Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HITLRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{hitl_id}/approve": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Approve Hitl Request",
        "description": "Approve a pending HITL request and resume tool execution.\n\nCommit-Before-Execute Ordering:\n  1. Verify PENDING status and non-expired.\n  2. Update hitl_request to APPROVED and tool_request to APPROVED.\n  3. Write and commit the HITL_APPROVED audit log entry.\n  4. Execute mock tool handler (if registered).\n  5. Write and commit the TOOL_EXECUTED / TOOL_EXECUTION_FAILED audit log entry.",
        "operationId": "approve_hitl_request_api_v1_hitl__hitl_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hitl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hitl Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HITLReviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Review Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HITLRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hitl/{hitl_id}/deny": {
      "post": {
        "tags": [
          "hitl"
        ],
        "summary": "Deny Hitl Request",
        "description": "Deny a pending HITL request. No mock tool handler is executed.",
        "operationId": "deny_hitl_request_api_v1_hitl__hitl_id__deny_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hitl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hitl Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/HITLReviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Review Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HITLRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HITLRequestListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HITLRequestResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "HITLRequestListResponse",
        "description": "Paginated list of HITL requests."
      },
      "HITLRequestResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "tool_request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tool Request Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "tool_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Name"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "input_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Payload"
          },
          "output_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "tool_request_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "HITLRequestResponse",
        "description": "Full detail of a Human-in-the-Loop approval request."
      },
      "HITLReviewRequest": {
        "properties": {
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          }
        },
        "type": "object",
        "title": "HITLReviewRequest",
        "description": "Optional payload when approving or denying a HITL request."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}